Downgrade embedlite to fb71d63ef4e6fd52ddbff08d4bd69c45138f0561 revision Fixed problem with horizontal scrolling performance

+++ gfx/layers/ipc/Axis.cpp
@@ -145,7 +145,7 @@ void Axis::UpdateWithTouchAtDevicePoint(int32_t aPos, const TimeDuration& aTimeD
 
   // Keep last gMaxVelocityQueueSize or less velocities in the queue.
   mVelocityQueue.AppendElement(mVelocity);
-  if (mVelocityQueue.Length() > (uint32_t)gMaxVelocityQueueSize) {
+  if (mVelocityQueue.Length() > gMaxVelocityQueueSize) {
     mVelocityQueue.RemoveElementAt(0);
   }
 }
